% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rhat.R
\name{calc_rhat}
\alias{calc_rhat}
\title{R hat}
\usage{
calc_rhat(x)
}
\arguments{
\item{x}{is the list of chains}
}
\value{
a vector of rhats
}
\description{
This function calculates the rhat
of each parameter given the list of chains 
that walkr produces. Since this is just an 
internal function, I'll document it more later.
}
\examples{
x <- matrix(seq(1,12,1), ncol = 3, nrow = 4) 
calc_rhat(list(x,x,x))

}
