% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/complete_solution.R
\name{complete_solution}
\alias{complete_solution}
\title{Complete Solution}
\usage{
complete_solution(A, b)
}
\arguments{
\item{A}{is the lhs of the matrix equation \eqn{Ax=b}}

\item{b}{is the rhs of the matrix equation \eqn{Ax=b}}
}
\value{
a list object, with the first element($particular) as the particular solution
        and the the second element as a matrix with its columns containing
        the basis of the null space($homogeneous)
}
\description{
Given matrix equation \eqn{Aw=b}, find
the basis representation of the
complete solution (an affine transformation),
returning the homogeneous and particular solution
which describe the polytope in alpha-space
}

