% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diag_logbarrier.R
\name{diag_logbarrier}
\alias{diag_logbarrier}
\title{diagonalized log-barrier}
\usage{
diag_logbarrier(A, b, x)
}
\arguments{
\item{A}{is the lhs of \eqn{Ax <= b}}

\item{b}{is the rhs of \eqn{Ax <= b}}

\item{x}{is the starting point}
}
\value{
a diagonalized matrix of 1/logbarrier function
}
\description{
Computes the diagonal matrix from the vector obtained
from 1 / the log-barrier function
}

