% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dikin_ellipsoid.R
\name{dikin_ellipsoid}
\alias{dikin_ellipsoid}
\title{Dikin Ellipsoid}
\usage{
dikin_ellipsoid(A, b, x0, r, z)
}
\arguments{
\item{A}{is the lhs of \eqn{Ax <= b}}

\item{b}{is the rhs of \eqn{Ax <= b}}

\item{x0}{is the starting point}

\item{r}{is the radius of the ellipsoid}

\item{z}{is the point we want to test}
}
\value{
a boolean indicating whether z is in the ellipsoid
}
\description{
Given specific parameters,
this function checks whether a point
z is in the specified Dikin Ellipsoid
centered at x0
}

