% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hessian_logbarrier.R
\name{hessian_logbarrier}
\alias{hessian_logbarrier}
\title{Hessian of Log Barrier}
\usage{
hessian_logbarrier(A, b, x)
}
\arguments{
\item{A}{is the lhs of \eqn{Ax \le b}}

\item{b}{is the rhs of \eqn{Ax \le b}}

\item{x}{is the current point we're at}
}
\value{
the Hessian matrix for the specific parameters
}
\description{
At point x0, this function computes the Hessian
of the log barrier function for the
convex polytope that is defined by \eqn{Ax \le b}.
}

