% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hit_and_run.R
\name{hit_and_run}
\alias{hit_and_run}
\title{Hit and Run}
\usage{
hit_and_run(A, b, x0, n)
}
\arguments{
\item{A}{is the lhs of the inequation Ax <= b}

\item{b}{is the rhs of the inequation Ax <= b}

\item{x0}{is the starting point, must be within the convex polytope}

\item{n}{is the number of points we want to sample}
}
\value{
a matrix, with the columns as points sampled
}
\description{
Given Ax <= b, this function samples from the convex polytope
using the hit-and-run algorithm
}

