% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/optimizedDikin.R
\name{optim_dikin_walk}
\alias{optim_dikin_walk}
\title{Optimized Dikin Walk}
\usage{
optim_dikin_walk(A, b, x0 = NULL, n, r = 1, analytic_center = T)
}
\arguments{
\item{A}{is the lhs of Ax <= b}

\item{b}{is the rhs of Ax <= b}

\item{x0}{is the starting point}

\item{n}{is the number of points we want}

\item{r}{is the radius of the ellipsoid (1 by default)}

\item{analytic_center}{is a boolean indicating whether we want to
      start at the analytic center of the polytope}
}
\value{
a number of sampled points that satisfy Ax <= b (matrix object,
columns the points)
}
\description{
This function implements the Dikin Walk using the Hessian
of the Log barrier function. Note that a $r$ of 1 guarantees
that the ellipsoid generated won't leave our polytope $K$ (see
Theorems online)
}

