% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcppeigen_fcrossprod}
\alias{rcppeigen_fcrossprod}
\title{Fast Matrix Cross-Product}
\usage{
rcppeigen_fcrossprod(A, B)
}
\arguments{
\item{A}{is the first parameter in t(A) times B}

\item{B}{is the second parameter in t(A) times B}
}
\value{
matrix cross-product t(A) times B
}
\description{
Computes using RcppEigen the product of t(A) and B
}
\examples{
\dontrun{
rcppeigen_fcrossprod(A, B)
}
}

