% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcppeigen_fdet}
\alias{rcppeigen_fdet}
\title{Fast Matrix Determinant}
\usage{
rcppeigen_fdet(A)
}
\arguments{
\item{A}{is the matrix whose determinant calculated}
}
\value{
determinant of A
}
\description{
Computes using RcppEigen the determinant of A
}
\examples{
\dontrun{
rcppeigen_fdet(A)
}

}
