% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcppeigen_fprod}
\alias{rcppeigen_fprod}
\title{Fast Matrix Product}
\usage{
rcppeigen_fprod(A, B)
}
\arguments{
\item{A}{is the first parameter in A times B}

\item{B}{is the second parameter in A times B}
}
\value{
matrix product A times B
}
\description{
Computes using RcppEigen the product of A and B
}
\examples{
\dontrun{
rcppeigen_fprod(A, B)
}
}

