% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcppeigen_fsolve}
\alias{rcppeigen_fsolve}
\title{Fast Matrix Inverse}
\usage{
rcppeigen_fsolve(A)
}
\arguments{
\item{A}{is the matrix being inverted}
}
\value{
inverse of A
}
\description{
Computes using RcppEigen the inverse of A
}
\examples{
\dontrun{
rcppeigen_fsolve(A)
}
}

