% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcppeigen_ftcrossprod}
\alias{rcppeigen_ftcrossprod}
\title{Fast Matrix T-Cross-Product}
\usage{
rcppeigen_ftcrossprod(A, B)
}
\arguments{
\item{A}{is the first parameter in A times t(B)}

\item{B}{is the second parameter in A times t(B)}
}
\value{
matrix tcross-product A times t(B)
}
\description{
Computes using RcppEigen the product of A and t(B)
}
\examples{
\dontrun{
rcppeigen_ftcrossprod(A, B)
}

}
