% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcppeigen_ftrans}
\alias{rcppeigen_ftrans}
\title{Fast Matrix Transpose}
\usage{
rcppeigen_ftrans(A)
}
\arguments{
\item{A}{is the matrix being transposed}
}
\value{
transpose of A
}
\description{
Computes using RcppEigen transposed of A
}
\examples{
\dontrun{
rcppeigen_ftrans(A)
}
}

