% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vis_sampling.R
\name{vis_sampling}
\alias{vis_sampling}
\title{Visualise Sampling}
\usage{
vis_sampling(x, chains)
}
\arguments{
\item{x}{is the set of points sampled}
}
\value{
a shiny interface that display the set of sampled points
}
\description{
This function takes in a set of sample points and diagnoses them. The final
results are displayed in a Shiny app. The app contains the confidence
interval of each dimension's coordiates across the whole sets of points,
and draws a graph about the distribution of all points.
}

