% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dikin_walk.R
\name{dikin_walk}
\alias{dikin_walk}
\title{Dikin Walk}
\usage{
dikin_walk(A, b, x0, n, r = 1)
}
\arguments{
\item{A}{is the lhs of \eqn{Ax \le b}}

\item{b}{is the rhs of \eqn{Ax \le b}}

\item{x0}{is the starting point}

\item{n}{is the number of points we want to sample}

\item{r}{is the radius of the Dikin ellipsoid (1 by default)}
}
\value{
a number of sampled points that satisfy Ax <= b (matrix object,
columns the points)
}
\description{
This function implements the Dikin Walk when given a
convex polytope defined by \eqn{Ax \le b} and a starting
point x0.
Note that a $r$ of 1 guarantees
that the ellipsoid generated won't leave our polytope $K$ (see
Theorems online)
}

