% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hit_and_run.R
\name{hit_and_run}
\alias{hit_and_run}
\title{Hit and Run}
\usage{
hit_and_run(A, b, x0, points, thin = 1, burn = 0)
}
\arguments{
\item{A}{is the lhs of the matrix equation A}

\item{b}{is the rhs of the matrix equation b}

\item{points}{is the number of points we want to sample}

\item{thin}{-- every thin-th point is taken}

\item{burn}{the first burn number of points are deleted}
}
\value{
a number of sampled points that satisfy Ax <= b (matrix object,
       columns the points)
}
\description{
This function provides a wrapper
for the har function of the
hit-and-run package
}

