% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vis_sampling.R
\name{vis_sampling}
\alias{vis_sampling}
\title{Visualise Sampling}
\usage{
vis_sampling(x, chains)
}
\arguments{
\item{x}{is the set of points sampled, with its columns as the sampled points.
If multiple chains are present, then the columns should be ordered
such that each chain follow each other.}

\item{chains}{is the number of chains that x contains}
}
\value{
a shiny interface that display the diagnostics of the MCMC random walk
}
\description{
This function takes in a set of sample points and diagnoses them using
the shinyStan interface. The app contains the confidence
interval of each dimension's coordiates across the whole sets of points,
Gelman-Rubin statistics, trace-plots, and other diagnostic tools for
examining convergence.
}

