% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/walkr.R
\name{walkr}
\alias{walkr}
\title{walkr}
\usage{
walkr(A, b, n, method)
}
\arguments{
\item{A}{is the lhs of the matrix equation A}

\item{b}{is the rhs of the matrix equation b}

\item{n}{is the number of points we want to sample}

\item{method}{is the MCMC sampling method. Please enter "hit-and-run" or "dikin"}
}
\value{
a list object. The first element is a matrix with its columns as the sampled
        points. The second element is the diagnostics of the MCMC sampling.
}
\description{
Given Ax = b, sample points from the intersection of Ax = b
with the n-simplex (\eqn{\sum(x) == 1}, \eqn{x_i >= 0}). The
current MCMC sampling methods supported are "hit-and-run" and
"dikin"
}

